indent and new apply_options function
