+++ include/key_txts.h
@@ -1,0 +1,42 @@
+// Copyright 2024 Rasmus Edgar
+static char KEY_TILE[] = KEY_TILE_TXT,
+        KEY_HGRID[] = KEY_HGRID_TXT,
+        KEY_SPIRAL[] = KEY_SPIRAL_TXT,
+        KEY_DWINDLE[] = KEY_DWINDLE_TXT,
+        KEY_SWITCHF[] = KEY_SWITCHF_TXT,
+        KEY_SWITCHB[] = KEY_SWITCHB_TXT,
+        KEY_INCTOPGAP[] = KEY_INCTOPGAP_TXT,
+        KEY_INCBOTTOMGAP[] = KEY_INCBOTTOMGAP_TXT,
+        KEY_INCLEFTGAP[] = KEY_INCLEFTGAP_TXT,
+        KEY_INCRIGHTGAP[] = KEY_INCRIGHTGAP_TXT,
+        KEY_DECTOPGAP[] = KEY_DECTOPGAP_TXT,
+        KEY_DECBOTTOMGAP[] = KEY_DECBOTTOMGAP_TXT,
+        KEY_DECLEFTGAP[] = KEY_DECLEFTGAP_TXT,
+        KEY_DECRIGHTGAP[] = KEY_DECRIGHTGAP_TXT,
+        KEY_INCALLGAPS[] = KEY_INCALLGAPS_TXT,
+        KEY_DECALLGAPS[] = KEY_DECALLGAPS_TXT,
+        KEY_TILE_OFF[] = KEY_TILE_OFF_TXT,
+        KEY_CMD_0[] = KEY_CMD_0_TXT,
+        KEY_CMD_1[] = KEY_CMD_1_TXT,
+        KEY_CMD_2[] = KEY_CMD_2_TXT,
+        KEY_CMD_3[] = KEY_CMD_3_TXT,
+        KEY_CMD_4[] = KEY_CMD_4_TXT,
+        KEY_CMD_5[] = KEY_CMD_5_TXT,
+        KEY_CMD_6[] = KEY_CMD_6_TXT,
+        KEY_CMD_7[] = KEY_CMD_7_TXT,
+        KEY_CMD_8[] = KEY_CMD_8_TXT,
+        KEY_CMD_9[] = KEY_CMD_9_TXT,
+        KEY_WS_0[] = KEY_WS_0_TXT,
+        KEY_WS_1[] = KEY_WS_1_TXT,
+        KEY_WS_2[] = KEY_WS_2_TXT,
+        KEY_WS_3[] = KEY_WS_3_TXT,
+        KEY_WS_4[] = KEY_WS_4_TXT,
+        KEY_WS_5[] = KEY_WS_5_TXT,
+        KEY_CWS_0[] = KEY_CWS_0_TXT,
+        KEY_CWS_1[] = KEY_CWS_1_TXT,
+        KEY_CWS_2[] = KEY_CWS_2_TXT,
+        KEY_CWS_3[] = KEY_CWS_3_TXT,
+        KEY_CWS_4[] = KEY_CWS_4_TXT,
+        KEY_CWS_5[] = KEY_CWS_5_TXT,
+        KEY_CXM_EXIT[] = KEY_CXM_EXIT_TXT,
+        KEY_TAB_NEXT[] = KEY_TAB_NEXT_TXT;
