+++ include/opt_defs.h
@@ -1,0 +1,83 @@
+// Copyright 2024 Rasmus Edgar
+Opts defopts[] =
+{
+        { TYPE_TOPGAP, TOPGAP_ID, OPTTYPE_ID },
+        { TYPE_DEFAULT_TOPGAP, DEFAULT_TOPGAP_ID, OPTTYPE_ID },
+        { TYPE_BOTTOMGAP, BOTTOMGAP_ID, OPTTYPE_ID },
+        { TYPE_LEFTGAP, LEFTGAP_ID, OPTTYPE_ID },
+        { TYPE_RIGHTGAP, RIGHTGAP_ID, OPTTYPE_ID },
+        { TYPE_AUTO, AUTO_ID, OPTTYPE_ID },
+        { TYPE_AUTO_INTERVAL_DELAY, AUTO_INTERVAL_DELAY_ID, OPTTYPE_ID },
+        { TYPE_TILE_FACT, TILE_FACT_ID, OPTTYPE_ID },
+        { TYPE_GAP_CHANGE_VALUE, GAP_CHANGE_VALUE_ID, OPTTYPE_ID },
+        { TYPE_BAR, BAR_ID, OPTTYPE_ID },
+        { TYPE_BAR_HIDE_EMPTY, BAR_HIDE_EMPTY_ID, OPTTYPE_ID },
+        { TYPE_BAR_BG_COL, BAR_BG_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPW_COL, BAR_FPW_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPW_COL, BAR_BPW_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPCURW_COL, BAR_FPCURW_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPCURW_COL, BAR_BPCURW_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPTM_COL, BAR_FPTM_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPTM_COL, BAR_BPTM_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPTI_COL, BAR_FPTI_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPTI_COL, BAR_BPTI_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPSEP_ONE_COL, BAR_FPSEP_ONE_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPSEP_ONE_COL, BAR_BPSEP_ONE_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_FPSEP_TWO_COL, BAR_FPSEP_TWO_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_BPSEP_TWO_COL, BAR_BPSEP_TWO_COL_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS0, BAR_TEXT_WS0_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS1, BAR_TEXT_WS1_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS2, BAR_TEXT_WS2_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS3, BAR_TEXT_WS3_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS4, BAR_TEXT_WS4_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_WS5, BAR_TEXT_WS5_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_TILE, BAR_TEXT_TILE_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_GRID, BAR_TEXT_GRID_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_DWINDLE, BAR_TEXT_DWINDLE_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_SPIRAL, BAR_TEXT_SPIRAL_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_SEP_1, BAR_TEXT_SEP_1_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_SEP_2, BAR_TEXT_SEP_2_ID, OPTTYPE_ID },
+        { TYPE_BAR_TEXT_SPACE, BAR_TEXT_SPACE_ID, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_0, EXCL_WTYPE_ID_0, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_1, EXCL_WTYPE_ID_1, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_2, EXCL_WTYPE_ID_2, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_3, EXCL_WTYPE_ID_3, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_4, EXCL_WTYPE_ID_4, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_5, EXCL_WTYPE_ID_5, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_6, EXCL_WTYPE_ID_6, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_7, EXCL_WTYPE_ID_7, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_8, EXCL_WTYPE_ID_8, OPTTYPE_ID },
+        { TYPE_EXCL_WTYPE_9, EXCL_WTYPE_ID_9, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_0, INCL_WTYPE_ID_0, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_1, INCL_WTYPE_ID_1, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_2, INCL_WTYPE_ID_2, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_3, INCL_WTYPE_ID_3, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_4, INCL_WTYPE_ID_4, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_5, INCL_WTYPE_ID_5, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_6, INCL_WTYPE_ID_6, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_7, INCL_WTYPE_ID_7, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_8, INCL_WTYPE_ID_8, OPTTYPE_ID },
+        { TYPE_INCL_WTYPE_9, INCL_WTYPE_ID_9, OPTTYPE_ID },
+        { TYPE_CONLINE_0, CONLINE_ID_0, OPTTYPE_ID },
+        { TYPE_CONLINE_1, CONLINE_ID_1, OPTTYPE_ID },
+        { TYPE_CONLINE_2, CONLINE_ID_2, OPTTYPE_ID },
+        { TYPE_CONLINE_3, CONLINE_ID_3, OPTTYPE_ID },
+        { TYPE_CONLINE_4, CONLINE_ID_4, OPTTYPE_ID },
+        { TYPE_CONLINE_5, CONLINE_ID_5, OPTTYPE_ID },
+        { TYPE_CONLINE_6, CONLINE_ID_6, OPTTYPE_ID },
+        { TYPE_CONLINE_7, CONLINE_ID_7, OPTTYPE_ID },
+        { TYPE_CONLINE_8, CONLINE_ID_8, OPTTYPE_ID },
+        { TYPE_CONLINE_9, CONLINE_ID_9, OPTTYPE_ID },
+        { TYPE_CMD_0, CMD_ID_0, OPTTYPE_ID },
+        { TYPE_CMD_1, CMD_ID_1, OPTTYPE_ID },
+        { TYPE_CMD_2, CMD_ID_2, OPTTYPE_ID },
+        { TYPE_CMD_3, CMD_ID_3, OPTTYPE_ID },
+        { TYPE_CMD_4, CMD_ID_4, OPTTYPE_ID },
+        { TYPE_CMD_5, CMD_ID_5, OPTTYPE_ID },
+        { TYPE_CMD_6, CMD_ID_6, OPTTYPE_ID },
+        { TYPE_CMD_7, CMD_ID_7, OPTTYPE_ID },
+        { TYPE_CMD_8, CMD_ID_8, OPTTYPE_ID },
+        { TYPE_CMD_9, CMD_ID_9, OPTTYPE_ID },
+        { TYPE_INFO_OFF, INFO_OFF_ID, OPTTYPE_ID },
+        { TYPE_VWS_ON, VWS_ON_ID, OPTTYPE_ID },
+};
