+++ src/commodity.c
@@ -1,422 +1,105 @@
-// This is a personal academic project. Dear PVS-Studio, please check it.
-// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
-// Copyright 2021 Rasmus Edgar
+// Copyright 2024 Rasmus Edgar
 #include "../include/dintwm_shared.h"
 #include "../include/commodity.h"
-
-static unsigned char COMMODITY_NAME[] = "DintWM commodity";
-static unsigned char COMMODITY_TITLE[] = "DintWM - a tiling window manager";
-static unsigned char COMMODITY_DESC[] = "To change hotkeys edit tooltypes";
-
-int fact = TILE_FACT_DEF;
-int gap_change_value = GAP_CHANGE_VALUE_DEF;
-static short info_on = TRUE;
-
-static unsigned long mainsig, wakeupsigs, subsig;
-static struct Task *maintask = NULL, *subtask = NULL;
-static unsigned char subactionchkname[] = "CXM_window_state_checker";
-static short first_run = TRUE;
-static struct Window *awin_comp;
-static short running = TRUE;
-static short autotile = FALSE;
-
-static char KEY_TILE[] = KEY_TILE_TXT,
-	KEY_HGRID[] = KEY_HGRID_TXT,
-	KEY_SPIRAL[] = KEY_SPIRAL_TXT,
-	KEY_DWINDLE[] = KEY_DWINDLE_TXT,
-	KEY_SWITCHF[] = KEY_SWITCHF_TXT,
-	KEY_SWITCHB[] = KEY_SWITCHB_TXT,
-	KEY_INCTOPGAP[] = KEY_INCTOPGAP_TXT,
-	KEY_INCBOTTOMGAP[] = KEY_INCBOTTOMGAP_TXT,
-	KEY_INCLEFTGAP[] = KEY_INCLEFTGAP_TXT,
-	KEY_INCRIGHTGAP[] = KEY_INCRIGHTGAP_TXT,
-	KEY_DECTOPGAP[] = KEY_DECTOPGAP_TXT,
-	KEY_DECBOTTOMGAP[] = KEY_DECBOTTOMGAP_TXT,
-	KEY_DECLEFTGAP[] = KEY_DECLEFTGAP_TXT,
-	KEY_DECRIGHTGAP[] = KEY_DECRIGHTGAP_TXT,
-	KEY_INCALLGAPS[] = KEY_INCALLGAPS_TXT,
-	KEY_DECALLGAPS[] = KEY_DECALLGAPS_TXT,
-	KEY_TILE_OFF[] = KEY_TILE_OFF_TXT,
-	KEY_CMD_0[] = KEY_CMD_0_TXT,
-	KEY_CMD_1[] = KEY_CMD_1_TXT,
-	KEY_CMD_2[] = KEY_CMD_2_TXT,
-	KEY_CMD_3[] = KEY_CMD_3_TXT,
-	KEY_CMD_4[] = KEY_CMD_4_TXT,
-	KEY_CMD_5[] = KEY_CMD_5_TXT,
-	KEY_CMD_6[] = KEY_CMD_6_TXT,
-	KEY_CMD_7[] = KEY_CMD_7_TXT,
-	KEY_CMD_8[] = KEY_CMD_8_TXT,
-	KEY_CMD_9[] = KEY_CMD_9_TXT,
-	KEY_WS_0[] = KEY_WS_0_TXT,
-	KEY_WS_1[] = KEY_WS_1_TXT,
-	KEY_WS_2[] = KEY_WS_2_TXT,
-	KEY_WS_3[] = KEY_WS_3_TXT,
-	KEY_WS_4[] = KEY_WS_4_TXT,
-	KEY_WS_5[] = KEY_WS_5_TXT,
-	KEY_CWS_0[] = KEY_CWS_0_TXT,
-	KEY_CWS_1[] = KEY_CWS_1_TXT,
-	KEY_CWS_2[] = KEY_CWS_2_TXT,
-	KEY_CWS_3[] = KEY_CWS_3_TXT,
-	KEY_CWS_4[] = KEY_CWS_4_TXT,
-	KEY_CWS_5[] = KEY_CWS_5_TXT,
-	KEY_CXM_EXIT[] = KEY_CXM_EXIT_TXT,
-	KEY_TAB_NEXT[] = KEY_TAB_NEXT_TXT;
-
-Keys defkeys[] = {
-	{ TYPE_TILE, KEY_TILE, KEYTYPE_ID, tile, {0} },
-	{ TYPE_HGRID, KEY_HGRID, KEYTYPE_ID, hgrid, {0} },
-	{ TYPE_SPIRAL, KEY_SPIRAL, KEYTYPE_ID, fibonacci, { .i = 0} },
-	{ TYPE_DWINDLE, KEY_DWINDLE, KEYTYPE_ID, fibonacci, { .i = 1 } },
-	{ TYPE_SWITCHF, KEY_SWITCHF, KEYTYPE_ID, switcher, { .i = 1 } },
-	{ TYPE_SWITCHB, KEY_SWITCHB, KEYTYPE_ID, switcher, { .i = 0 } },
-	{ TYPE_INCTOPGAP, KEY_INCTOPGAP, KEYTYPE_ID, changegaps, {.i = TOPGAP_ID} },
-	{ TYPE_INCBOTTOMGAP, KEY_INCBOTTOMGAP, KEYTYPE_ID, changegaps, {.i = BOTTOMGAP_ID} },
-	{ TYPE_INCLEFTGAP, KEY_INCLEFTGAP, KEYTYPE_ID, changegaps, {.i = LEFTGAP_ID} },
-	{ TYPE_INCRIGHTGAP, KEY_INCRIGHTGAP, KEYTYPE_ID, changegaps, {.i = RIGHTGAP_ID} },
-	{ TYPE_DECTOPGAP, KEY_DECTOPGAP, KEYTYPE_ID, changegaps, {.i = -TOPGAP_ID} },
-	{ TYPE_DECBOTTOMGAP, KEY_DECBOTTOMGAP, KEYTYPE_ID, changegaps, {.i = -BOTTOMGAP_ID} },
-	{ TYPE_DECLEFTGAP, KEY_DECLEFTGAP, KEYTYPE_ID, changegaps, {.i = -LEFTGAP_ID} },
-	{ TYPE_DECRIGHTGAP, KEY_DECRIGHTGAP, KEYTYPE_ID, changegaps, {.i = -RIGHTGAP_ID} },
-	{ TYPE_INCALLGAPS, KEY_INCALLGAPS, KEYTYPE_ID, changegaps, {.i = INCALLGAPS_ID} },
-	{ TYPE_DECALLGAPS, KEY_DECALLGAPS, KEYTYPE_ID, changegaps, {.i = DECALLGAPS_ID} },
-	{ TYPE_TILE_OFF, KEY_TILE_OFF, KEYTYPE_ID, tileoff, {0} },
-	{ TYPE_KEY_CMD_0, KEY_CMD_0, KEYTYPE_ID, docmd, { .i = CMD_ID_0 } },
-	{ TYPE_KEY_CMD_1, KEY_CMD_1, KEYTYPE_ID, docmd, { .i = CMD_ID_1 } },
-	{ TYPE_KEY_CMD_2, KEY_CMD_2, KEYTYPE_ID, docmd, { .i = CMD_ID_2 } },
-	{ TYPE_KEY_CMD_3, KEY_CMD_3, KEYTYPE_ID, docmd, { .i = CMD_ID_3 } },
-	{ TYPE_KEY_CMD_4, KEY_CMD_4, KEYTYPE_ID, docmd, { .i = CMD_ID_4 } },
-	{ TYPE_KEY_CMD_5, KEY_CMD_5, KEYTYPE_ID, docmd, { .i = CMD_ID_5 } },
-	{ TYPE_KEY_CMD_6, KEY_CMD_6, KEYTYPE_ID, docmd, { .i = CMD_ID_7 } },
-	{ TYPE_KEY_CMD_7, KEY_CMD_7, KEYTYPE_ID, docmd, { .i = CMD_ID_7 } },
-	{ TYPE_KEY_CMD_8, KEY_CMD_8, KEYTYPE_ID, docmd, { .i = CMD_ID_8 } },
-	{ TYPE_KEY_CMD_9, KEY_CMD_9, KEYTYPE_ID, docmd, { .i = CMD_ID_9 } },
-	{ TYPE_KEY_WS_0, KEY_WS_0, KEYTYPE_ID, changews, { .u = WS_0 } },
-	{ TYPE_KEY_WS_1, KEY_WS_1, KEYTYPE_ID, changews, { .u = WS_1 } },
-	{ TYPE_KEY_WS_2, KEY_WS_2, KEYTYPE_ID, changews, { .u = WS_2 } },
-	{ TYPE_KEY_WS_3, KEY_WS_3, KEYTYPE_ID, changews, { .u = WS_3 } },
-	{ TYPE_KEY_WS_4, KEY_WS_4, KEYTYPE_ID, changews, { .u = WS_4 } },
-	{ TYPE_KEY_WS_5, KEY_WS_5, KEYTYPE_ID, changews, { .u = WS_5 } },
-	{ TYPE_KEY_CWS_0, KEY_CWS_0, KEYTYPE_ID, movetows, { .u = WS_0 } },
-	{ TYPE_KEY_CWS_1, KEY_CWS_1, KEYTYPE_ID, movetows, { .u = WS_1 } },
-	{ TYPE_KEY_CWS_2, KEY_CWS_2, KEYTYPE_ID, movetows, { .u = WS_2 } },
-	{ TYPE_KEY_CWS_3, KEY_CWS_3, KEYTYPE_ID, movetows, { .u = WS_3 } },
-	{ TYPE_KEY_CWS_4, KEY_CWS_4, KEYTYPE_ID, movetows, { .u = WS_4 } },
-	{ TYPE_KEY_CWS_5, KEY_CWS_5, KEYTYPE_ID, movetows, { .u = WS_5 } },
-	{ TYPE_KEY_TAB_NEXT, KEY_TAB_NEXT, KEYTYPE_ID, tabnextwin, { .i = 0 } },
-	{ TYPE_KEY_CXM_EXIT, KEY_CXM_EXIT, KEYTYPE_ID, exit_cxm, {.s = FALSE} },
-};
-
-Opts defopts[] = {
-	{ TYPE_TOPGAP, TOPGAP_ID, OPTTYPE_ID },
-	{ TYPE_DEFAULT_TOPGAP, DEFAULT_TOPGAP_ID, OPTTYPE_ID },
-	{ TYPE_BOTTOMGAP, BOTTOMGAP_ID, OPTTYPE_ID },
-	{ TYPE_LEFTGAP, LEFTGAP_ID, OPTTYPE_ID },
-	{ TYPE_RIGHTGAP, RIGHTGAP_ID, OPTTYPE_ID },
-	{ TYPE_AUTO, AUTO_ID, OPTTYPE_ID },
-	{ TYPE_AUTO_INTERVAL_DELAY, AUTO_INTERVAL_DELAY_ID, OPTTYPE_ID },
-	{ TYPE_TILE_FACT, TILE_FACT_ID, OPTTYPE_ID },
-	{ TYPE_GAP_CHANGE_VALUE, GAP_CHANGE_VALUE_ID, OPTTYPE_ID },
-	{ TYPE_BAR, BAR_ID, OPTTYPE_ID },
-	{ TYPE_BAR_HIDE_EMPTY, BAR_HIDE_EMPTY_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BG_COL, BAR_BG_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPW_COL, BAR_FPW_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPW_COL, BAR_BPW_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPCURW_COL, BAR_FPCURW_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPCURW_COL, BAR_BPCURW_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPTM_COL, BAR_FPTM_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPTM_COL, BAR_BPTM_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPTI_COL, BAR_FPTI_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPTI_COL, BAR_BPTI_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPSEP_ONE_COL, BAR_FPSEP_ONE_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPSEP_ONE_COL, BAR_BPSEP_ONE_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_FPSEP_TWO_COL, BAR_FPSEP_TWO_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_BPSEP_TWO_COL, BAR_BPSEP_TWO_COL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS0, BAR_TEXT_WS0_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS1, BAR_TEXT_WS1_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS2, BAR_TEXT_WS2_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS3, BAR_TEXT_WS3_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS4, BAR_TEXT_WS4_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_WS5, BAR_TEXT_WS5_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_TILE, BAR_TEXT_TILE_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_GRID, BAR_TEXT_GRID_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_DWINDLE, BAR_TEXT_DWINDLE_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_SPIRAL, BAR_TEXT_SPIRAL_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_SEP_1, BAR_TEXT_SEP_1_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_SEP_2, BAR_TEXT_SEP_2_ID, OPTTYPE_ID },
-	{ TYPE_BAR_TEXT_SPACE, BAR_TEXT_SPACE_ID, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_0, EXCL_WTYPE_ID_0, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_1, EXCL_WTYPE_ID_1, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_2, EXCL_WTYPE_ID_2, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_3, EXCL_WTYPE_ID_3, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_4, EXCL_WTYPE_ID_4, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_5, EXCL_WTYPE_ID_5, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_6, EXCL_WTYPE_ID_6, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_7, EXCL_WTYPE_ID_7, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_8, EXCL_WTYPE_ID_8, OPTTYPE_ID },
-	{ TYPE_EXCL_WTYPE_9, EXCL_WTYPE_ID_9, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_0, INCL_WTYPE_ID_0, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_1, INCL_WTYPE_ID_1, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_2, INCL_WTYPE_ID_2, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_3, INCL_WTYPE_ID_3, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_4, INCL_WTYPE_ID_4, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_5, INCL_WTYPE_ID_5, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_6, INCL_WTYPE_ID_6, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_7, INCL_WTYPE_ID_7, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_8, INCL_WTYPE_ID_8, OPTTYPE_ID },
-	{ TYPE_INCL_WTYPE_9, INCL_WTYPE_ID_9, OPTTYPE_ID },
-	{ TYPE_CONLINE_0, CONLINE_ID_0, OPTTYPE_ID },
-	{ TYPE_CONLINE_1, CONLINE_ID_1, OPTTYPE_ID },
-	{ TYPE_CONLINE_2, CONLINE_ID_2, OPTTYPE_ID },
-	{ TYPE_CONLINE_3, CONLINE_ID_3, OPTTYPE_ID },
-	{ TYPE_CONLINE_4, CONLINE_ID_4, OPTTYPE_ID },
-	{ TYPE_CONLINE_5, CONLINE_ID_5, OPTTYPE_ID },
-	{ TYPE_CONLINE_6, CONLINE_ID_6, OPTTYPE_ID },
-	{ TYPE_CONLINE_7, CONLINE_ID_7, OPTTYPE_ID },
-	{ TYPE_CONLINE_8, CONLINE_ID_8, OPTTYPE_ID },
-	{ TYPE_CONLINE_9, CONLINE_ID_9, OPTTYPE_ID },
-	{ TYPE_CMD_0, CMD_ID_0, OPTTYPE_ID },
-	{ TYPE_CMD_1, CMD_ID_1, OPTTYPE_ID },
-	{ TYPE_CMD_2, CMD_ID_2, OPTTYPE_ID },
-	{ TYPE_CMD_3, CMD_ID_3, OPTTYPE_ID },
-	{ TYPE_CMD_4, CMD_ID_4, OPTTYPE_ID },
-	{ TYPE_CMD_5, CMD_ID_5, OPTTYPE_ID },
-	{ TYPE_CMD_6, CMD_ID_6, OPTTYPE_ID },
-	{ TYPE_CMD_7, CMD_ID_7, OPTTYPE_ID },
-	{ TYPE_CMD_8, CMD_ID_8, OPTTYPE_ID },
-	{ TYPE_CMD_9, CMD_ID_9, OPTTYPE_ID },
-	{ TYPE_INFO_OFF, INFO_OFF_ID, OPTTYPE_ID },
-	{ TYPE_VWS_ON, VWS_ON_ID, OPTTYPE_ID },
-};
-
-Ostore cmds[] = {0};
-Ostore cons[] = {0};
-Ostore incls[] = {0};
-Ostore excls[] = {0};
-
-static struct NewBroker MyBroker =
-{
-        NB_VERSION,
-        COMMODITY_NAME,
-        COMMODITY_TITLE,
-        COMMODITY_DESC,
-        NBU_UNIQUE | NBU_NOTIFY,
-        0,
-        0,
-        0,
-        0
-};
+// initialize commodity options and keycombos
+#include "../include/key_txts.h"
+#include "../include/key_defs.h"
+#include "../include/opt_defs.h"
 
 static short attachtooltypes(CxObj *broker, struct MsgPort *port, struct DiskObject *diskobj)
 {
 	short rc = TRUE;
 	size_t keyarrsize;
 	size_t optarrsize;
-	struct Popkeys* keys;
+	struct Popkeys *keys;
 	const char *bar_default_text[BAR_LAST_TEXT];
-        bar_default_text[ws_zero] = DEF_BAR_TEXT_WS_ZERO;
-        bar_default_text[ws_one] = DEF_BAR_TEXT_WS_ONE;
-        bar_default_text[ws_two] = DEF_BAR_TEXT_WS_TWO;
-        bar_default_text[ws_three] = DEF_BAR_TEXT_WS_THREE;
-        bar_default_text[ws_four] = DEF_BAR_TEXT_WS_FOUR;
-        bar_default_text[ws_five] = DEF_BAR_TEXT_WS_FIVE;
-        bar_default_text[ws_wb] = DEF_BAR_TEXT_WS_WB;
-        bar_default_text[mode_tile] = DEF_BAR_TEXT_MODE_TILE;
-        bar_default_text[mode_grid] = DEF_BAR_TEXT_MODE_GRID;
-        bar_default_text[mode_dwindle] = DEF_BAR_TEXT_MODE_DWINDLE;
-        bar_default_text[mode_spiral]= DEF_BAR_TEXT_MODE_SPIRAL;
-        bar_default_text[sep_one]= DEF_BAR_TEXT_SEP_ONE;
-        bar_default_text[sep_two]= DEF_BAR_TEXT_SEP_TWO;
-        bar_default_text[space]= DEF_BAR_TEXT_SPACE;
-        bar_default_text[err]= DEF_BAR_TEXT_ERR;
+	bar_default_text[ws_zero] = DEF_BAR_TEXT_WS_ZERO;
+	bar_default_text[ws_one] = DEF_BAR_TEXT_WS_ONE;
+	bar_default_text[ws_two] = DEF_BAR_TEXT_WS_TWO;
+	bar_default_text[ws_three] = DEF_BAR_TEXT_WS_THREE;
+	bar_default_text[ws_four] = DEF_BAR_TEXT_WS_FOUR;
+	bar_default_text[ws_five] = DEF_BAR_TEXT_WS_FIVE;
+	bar_default_text[ws_wb] = DEF_BAR_TEXT_WS_WB;
+	bar_default_text[mode_tile] = DEF_BAR_TEXT_MODE_TILE;
+	bar_default_text[mode_grid] = DEF_BAR_TEXT_MODE_GRID;
+	bar_default_text[mode_dwindle] = DEF_BAR_TEXT_MODE_DWINDLE;
+	bar_default_text[mode_spiral] = DEF_BAR_TEXT_MODE_SPIRAL;
+	bar_default_text[sep_one] = DEF_BAR_TEXT_SEP_ONE;
+	bar_default_text[sep_two] = DEF_BAR_TEXT_SEP_TWO;
+	bar_default_text[space] = DEF_BAR_TEXT_SPACE;
+	bar_default_text[err] = DEF_BAR_TEXT_ERR;
 
 	exclude_wtype = 0;
 	include_wtype = 0;
 
 	optarrsize = sizeof(defopts) / sizeof(*defopts);
-        keyarrsize = sizeof(defkeys) / sizeof(*defkeys);
+	keyarrsize = sizeof(defkeys) / sizeof(*defkeys);
 
-        bar_text = (Bar_Text *)malloc(BAR_LAST_TEXT * sizeof(*bar_text));
-        for (int i = 0; i < BAR_LAST_TEXT; i++) {
-                bar_text[i].text = malloc(TT_MAX_LENGTH * 2);
+	bar_text = (Bar_Text *) malloc(BAR_LAST_TEXT * sizeof(*bar_text));
+	for (int i = 0; i < BAR_LAST_TEXT; i++) {
+		bar_text[i].text = malloc(TT_MAX_LENGTH * 2);
 		bar_text[i].text[0] = nil;
-        }
+	}
 
-	for (size_t i = 0; i < optarrsize ; ++i) {
-       		const char *tt_optvalue = (char *)FindToolType((const unsigned char **)diskobj->do_ToolTypes, (const unsigned char *)defopts[i].optname);
-
-		if ((tt_optvalue) && ((strnlen(tt_optvalue, TT_MAX_LENGTH) < (size_t)TT_MAX_LENGTH))) {
-			if (defopts[i].cxint >= EXCL_WTYPE_ID_0 && defopts[i].cxint <= (WTYPE_MAX + EXCL_WTYPE_ID_0)) {
-				rc = alloc_opts(tt_optvalue, excls, i, EXCL_WTYPE_ID_0);
-				if (exclude_wtype == 0) {
-					exclude_wtype = 1;
-				}
-			}
-
-			if (defopts[i].cxint >= INCL_WTYPE_ID_0 && defopts[i].cxint <= (WTYPE_MAX + INCL_WTYPE_ID_0)) {
-				rc = alloc_opts(tt_optvalue, incls, i, INCL_WTYPE_ID_0);
-				if (include_wtype == 0) {
-					include_wtype = 1;
-				}
-			}
-
-			if (defopts[i].cxint >= CONLINE_ID_0 && defopts[i].cxint <= (CMD_MAX + CONLINE_ID_0)) {
-				rc = alloc_opts(tt_optvalue, cons, i, CONLINE_ID_0);
-			}
-
-			if (defopts[i].cxint >= CMD_ID_0 && defopts[i].cxint <= (CMD_MAX + CMD_ID_0)) {
-				rc = alloc_opts(tt_optvalue, cmds, i, CMD_ID_0);
-			}
-
-
-			switch (defopts[i].cxint) {
-				case TOPGAP_ID:
-					topgap = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case DEFAULT_TOPGAP_ID:
-					topgap = calcgap();
-					break;
-				case BOTTOMGAP_ID:
-					bottomgap = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case LEFTGAP_ID:
-					leftgap = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case RIGHTGAP_ID:
-					rightgap = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case GAP_CHANGE_VALUE_ID:
-					gap_change_value = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_ID:
-					bar_on = TRUE;
-					break;
-				case BAR_BG_COL_ID:
-					bar_color[bg].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPW_COL_ID:
-					bar_color[fp_ws].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPW_COL_ID:
-					bar_color[bp_ws].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPCURW_COL_ID:
-					bar_color[fp_cur].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPCURW_COL_ID:
-					bar_color[bp_cur].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPTM_COL_ID:
-					bar_color[fp_tm].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPTM_COL_ID:
-					bar_color[bp_tm].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPTI_COL_ID:
-					bar_color[fp_ti].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPTI_COL_ID:
-					bar_color[bp_ti].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPSEP_ONE_COL_ID:
-					bar_color[fp_sep_one].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPSEP_ONE_COL_ID:
-					bar_color[bp_sep_one].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_FPSEP_TWO_COL_ID:
-					bar_color[fp_sep_two].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_BPSEP_TWO_COL_ID:
-					bar_color[bp_sep_two].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case BAR_HIDE_EMPTY_ID:
-					hidewbar |= BAR_HIDE_ON;
-					break;
-				case BAR_TEXT_WS0_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_zero, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_WS1_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_one, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_WS2_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_two, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_WS3_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_three, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_WS4_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_four, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_WS5_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, ws_five, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_TILE_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, mode_tile, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_GRID_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, mode_grid, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_DWINDLE_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, mode_dwindle, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_SPIRAL_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, mode_spiral, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_SEP_1_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, sep_one, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_SEP_2_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, sep_two, tt_optvalue) : FALSE;
-					break;
-				case BAR_TEXT_SPACE_ID:
-					rc = rc == TRUE ? assign_bar_item(bar_text, space, tt_optvalue) : FALSE;
-					break;
-				case AUTO_ID:
-					autotile = TRUE;
-					break;
-				case AUTO_INTERVAL_DELAY_ID:
-					auto_interval = strtoul((const char *)tt_optvalue, (char **)NULL, 10);
-					break;
-				case TILE_FACT_ID:
-					fact = (int)strtol((const char*)tt_optvalue, (char **)NULL, 10);
-					break;
-				case INFO_OFF_ID:
-					info_on = FALSE;
-					break;
-				case VWS_ON_ID:
-					vws_on = TRUE;
-					break;
-				default:
-					// Do nothing
-					break;
-			}
+	for (size_t i = 0; i < optarrsize; ++i) {
+		const char *tt_optvalue =
+		    (char *)FindToolType((const unsigned char **)diskobj->do_ToolTypes, (const unsigned char *)defopts[i].optname);
+		rc = apply_options(defopts, tt_optvalue, i);
+		if (rc == FALSE) {
+			break;
 		}
 	}
 
 	if (rc == TRUE && bar_on == TRUE) {
-                for (int i = 0; i < BAR_LAST_TEXT; i++) {
+		for (int i = 0; i < BAR_LAST_TEXT; i++) {
 			if (bar_text[i].text[0] != nil) {
 				continue;
 			}
-                        if ((assign_bar_item(bar_text, i, bar_default_text[i]) == FALSE)) {
-                                return FALSE;
-                        }
-                }
+			if ((assign_bar_item(bar_text, i, bar_default_text[i]) == FALSE)) {
+				return FALSE;
+			}
+		}
 
 		// Bar color assignment
-		if (*bar_color[bg].color == nil) { bar_color[bg].color[0] = DEF_BAR_BG_COL; }
-		if (*bar_color[fp_ws].color == nil) { bar_color[fp_ws].color[0] = DEF_BAR_FP_WS_COL; }
-		if (*bar_color[bp_ws].color == nil) { bar_color[bp_ws].color[0] = DEF_BAR_BP_WS_COL; }
-		if (*bar_color[fp_cur].color == nil) { bar_color[fp_cur].color[0] = DEF_BAR_FP_CUR_COL; }
-		if (*bar_color[bp_cur].color == nil) { bar_color[bp_cur].color[0] = DEF_BAR_BP_CUR_COL; }
-		if (*bar_color[fp_tm].color == nil) { bar_color[fp_tm].color[0] = DEF_BAR_FP_TM_COL; }
-		if (*bar_color[bp_tm].color == nil) { bar_color[bp_tm].color[0] = DEF_BAR_BP_TM_COL; }
-		if (*bar_color[fp_ti].color == nil) { bar_color[fp_ti].color[0] = DEF_BAR_FP_TI_COL; }
-		if (*bar_color[bp_ti].color == nil) { bar_color[bp_ti].color[0] = DEF_BAR_BP_TI_COL; }
-		if (*bar_color[fp_sep_one].color == nil) { bar_color[fp_sep_one].color[0]  = DEF_BAR_FP_SEP_ONE_COL; }
-		if (*bar_color[bp_sep_one].color == nil) { bar_color[bp_sep_one].color[0] = DEF_BAR_BP_SEP_ONE_COL; }
-		if (*bar_color[fp_sep_two].color == nil) { bar_color[fp_sep_two].color[0] = DEF_BAR_FP_SEP_TWO_COL; }
-		if (*bar_color[bp_sep_two].color == nil) { bar_color[bp_sep_two].color[0] = DEF_BAR_BP_SEP_TWO_COL; }
+		if (*bar_color[bg].color == nil) {
+			bar_color[bg].color[0] = DEF_BAR_BG_COL;
+		}
+		if (*bar_color[fp_ws].color == nil) {
+			bar_color[fp_ws].color[0] = DEF_BAR_FP_WS_COL;
+		}
+		if (*bar_color[bp_ws].color == nil) {
+			bar_color[bp_ws].color[0] = DEF_BAR_BP_WS_COL;
+		}
+		if (*bar_color[fp_cur].color == nil) {
+			bar_color[fp_cur].color[0] = DEF_BAR_FP_CUR_COL;
+		}
+		if (*bar_color[bp_cur].color == nil) {
+			bar_color[bp_cur].color[0] = DEF_BAR_BP_CUR_COL;
+		}
+		if (*bar_color[fp_tm].color == nil) {
+			bar_color[fp_tm].color[0] = DEF_BAR_FP_TM_COL;
+		}
+		if (*bar_color[bp_tm].color == nil) {
+			bar_color[bp_tm].color[0] = DEF_BAR_BP_TM_COL;
+		}
+		if (*bar_color[fp_ti].color == nil) {
+			bar_color[fp_ti].color[0] = DEF_BAR_FP_TI_COL;
+		}
+		if (*bar_color[bp_ti].color == nil) {
+			bar_color[bp_ti].color[0] = DEF_BAR_BP_TI_COL;
+		}
+		if (*bar_color[fp_sep_one].color == nil) {
+			bar_color[fp_sep_one].color[0] = DEF_BAR_FP_SEP_ONE_COL;
+		}
+		if (*bar_color[bp_sep_one].color == nil) {
+			bar_color[bp_sep_one].color[0] = DEF_BAR_BP_SEP_ONE_COL;
+		}
+		if (*bar_color[fp_sep_two].color == nil) {
+			bar_color[fp_sep_two].color[0] = DEF_BAR_FP_SEP_TWO_COL;
+		}
+		if (*bar_color[bp_sep_two].color == nil) {
+			bar_color[bp_sep_two].color[0] = DEF_BAR_BP_SEP_TWO_COL;
+		}
 	}
 
 	if (rc == FALSE) {
@@ -434,8 +117,9 @@ static short attachtooltypes(CxObj *broker, struct MsgPort *port, struct DiskObj
 	keys = malloc(sizeof(*keys) * keyarrsize);
 
 	if (keys != NULL) {
-		for (size_t i = 0; i < keyarrsize ; ++i) {
-			keys[i].rawcombo = (char *)FindToolType((const unsigned char **)diskobj->do_ToolTypes, (const unsigned char *)defkeys[i].optname);
+		for (size_t i = 0; i < keyarrsize; ++i) {
+			keys[i].rawcombo =
+			    (char *)FindToolType((const unsigned char **)diskobj->do_ToolTypes, (const unsigned char *)defkeys[i].optname);
 
 			if (keys[i].rawcombo == NULL) {
 				keys[i].rawcombo = defkeys[i].defaultval;
@@ -443,8 +127,7 @@ static short attachtooltypes(CxObj *broker, struct MsgPort *port, struct DiskObj
 
 			if (keys[i].rawcombo != NULL) {
 				CxObj *filter;
-				if ((filter = HotKey((const unsigned char *)keys[i].rawcombo, port, (long int)i)) != NULL)
-				{
+				if ((filter = HotKey((const unsigned char *)keys[i].rawcombo, port, (long int)i)) != NULL) {
 					AttachCxObj(broker, filter);
 
 					if (CxObjError(filter) != 0) {
@@ -464,10 +147,175 @@ static short attachtooltypes(CxObj *broker, struct MsgPort *port, struct DiskObj
 	return rc;
 }
 
+static short apply_options(Opts const *dopts, const char *tt_optvalue, size_t i)
+{
+	short rc = TRUE;
+	if ((tt_optvalue)
+	    && ((strnlen(tt_optvalue, TT_MAX_LENGTH) < (size_t)TT_MAX_LENGTH))) {
+		if (dopts[i].cxint >= EXCL_WTYPE_ID_0 && dopts[i].cxint <= (WTYPE_MAX + EXCL_WTYPE_ID_0)) {
+			rc = alloc_opts(tt_optvalue, excls, i, EXCL_WTYPE_ID_0);
+			if (exclude_wtype == 0) {
+				exclude_wtype = 1;
+			}
+		}
+
+		if (dopts[i].cxint >= INCL_WTYPE_ID_0 && dopts[i].cxint <= (WTYPE_MAX + INCL_WTYPE_ID_0)) {
+			rc = alloc_opts(tt_optvalue, incls, i, INCL_WTYPE_ID_0);
+			if (include_wtype == 0) {
+				include_wtype = 1;
+			}
+		}
+
+		if (dopts[i].cxint >= CONLINE_ID_0 && dopts[i].cxint <= (CMD_MAX + CONLINE_ID_0)) {
+			rc = alloc_opts(tt_optvalue, cons, i, CONLINE_ID_0);
+		}
+
+		if (dopts[i].cxint >= CMD_ID_0 && dopts[i].cxint <= (CMD_MAX + CMD_ID_0)) {
+			rc = alloc_opts(tt_optvalue, cmds, i, CMD_ID_0);
+		}
+
+		switch (dopts[i].cxint) {
+		case TOPGAP_ID:
+			topgap = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case DEFAULT_TOPGAP_ID:
+			topgap = calcgap();
+			break;
+		case BOTTOMGAP_ID:
+			bottomgap = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case LEFTGAP_ID:
+			leftgap = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case RIGHTGAP_ID:
+			rightgap = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case GAP_CHANGE_VALUE_ID:
+			gap_change_value = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_ID:
+			bar_on = TRUE;
+			break;
+		case BAR_BG_COL_ID:
+			bar_color[bg].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPW_COL_ID:
+			bar_color[fp_ws].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPW_COL_ID:
+			bar_color[bp_ws].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPCURW_COL_ID:
+			bar_color[fp_cur].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPCURW_COL_ID:
+			bar_color[bp_cur].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPTM_COL_ID:
+			bar_color[fp_tm].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPTM_COL_ID:
+			bar_color[bp_tm].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPTI_COL_ID:
+			bar_color[fp_ti].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPTI_COL_ID:
+			bar_color[bp_ti].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPSEP_ONE_COL_ID:
+			bar_color[fp_sep_one].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPSEP_ONE_COL_ID:
+			bar_color[bp_sep_one].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_FPSEP_TWO_COL_ID:
+			bar_color[fp_sep_two].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_BPSEP_TWO_COL_ID:
+			bar_color[bp_sep_two].color[0] = (unsigned char)strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case BAR_HIDE_EMPTY_ID:
+			hidewbar |= BAR_HIDE_ON;
+			break;
+		case BAR_TEXT_WS0_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_zero, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_WS1_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_one, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_WS2_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_two, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_WS3_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_three, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_WS4_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_four, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_WS5_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, ws_five, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_TILE_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, mode_tile, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_GRID_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, mode_grid, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_DWINDLE_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, mode_dwindle, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_SPIRAL_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, mode_spiral, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_SEP_1_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, sep_one, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_SEP_2_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, sep_two, tt_optvalue) : FALSE;
+			break;
+		case BAR_TEXT_SPACE_ID:
+			rc = rc == TRUE ? assign_bar_item(bar_text, space, tt_optvalue) : FALSE;
+			break;
+		case AUTO_ID:
+			autotile = TRUE;
+			break;
+		case AUTO_INTERVAL_DELAY_ID:
+			auto_interval = strtoul((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case TILE_FACT_ID:
+			fact = (int)strtol((const char *)tt_optvalue, (char **)NULL, 10);
+			break;
+		case INFO_OFF_ID:
+			info_on = FALSE;
+			break;
+		case VWS_ON_ID:
+			vws_on = TRUE;
+			break;
+		default:
+			// Do nothing
+			break;
+		}
+	}
+	return rc;
+}
+
 void init_cmo(struct Cmo *cmo)
 {
-	unsigned char iconlib[] = "icon.library";
-	unsigned char diskobjname[] = "PROGDIR:dintwm";
+	static unsigned char COMMODITY_NAME[] = "DintWM commodity";
+	static unsigned char COMMODITY_TITLE[] = "DintWM - a tiling window manager";
+	static unsigned char COMMODITY_DESC[] = "To change hotkeys edit tooltypes";
+	static struct NewBroker MyBroker = {
+		NB_VERSION,
+		COMMODITY_NAME,
+		COMMODITY_TITLE,
+		COMMODITY_DESC,
+		NBU_UNIQUE | NBU_NOTIFY,
+		0,
+		0,
+		0,
+		0
+	};
 	enum {
 		mpfail,
 		iconfail,
@@ -478,6 +326,9 @@ void init_cmo(struct Cmo *cmo)
 		attfail,
 		acxofail
 	};
+	unsigned char iconlib[] = "icon.library";
+	unsigned char diskobjname[] = "PROGDIR:dintwm";
+
 	cmo->mp = NULL;
 	cmo->mp = CreateMsgPort();
 	if (cmo->mp == NULL) {
@@ -525,7 +376,6 @@ short int commo(void)
 	init_cmo(&cmo);
 	auto_interval = (unsigned long)AUTO_INTERVAL_DELAY_DEF;
 	size_t cmo_arr_length = sizeof(cmo.failarr) / sizeof(cmo.failarr[0]);
-	//printf("%s\n", cmo_fail_msgs[7]);
 
 	for (size_t i = 0; i < cmo_arr_length; i++) {
 		if (cmo.failarr[i] == -1) {
@@ -539,25 +389,20 @@ short int commo(void)
 	subsig = 1UL << (unsigned long)cmo.subsignum;
 	maintask = FindTask(NULL);
 
-	if (running == TRUE)
-	{
+	if (running == TRUE) {
 		CxMsg *msg;
 		CloseLibrary(cmo.iconbase);
 		cmo.iconbase = NULL;
 		FreeDiskObject(cmo.diskobj);
 		cmo.diskobj = NULL;
-
-		/*if (ActivateCxObj(cmo.broker, 1) != 0) {
-			cleanup(&cmo);
-			return EXIT_FAILURE;
-		}*/
+		static unsigned char subactionchkname[] = "CXM_window_state_checker";
 
 		// Muting GCC warning here. Following official Amiga CreateTask example
-		#pragma GCC diagnostic push
-		#pragma GCC diagnostic ignored "-Wpedantic"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wpedantic"
 		subtask = CreateTask(subactionchkname, 0L, (void *)subactionchk, 2048L);
-		#pragma GCC diagnostic pop
-		if(subtask == NULL) {
+#pragma GCC diagnostic pop
+		if (subtask == NULL) {
 			running = FALSE;
 		}
 
@@ -590,18 +435,16 @@ short int commo(void)
 				update_wbar();
 			}
 		}
-
 		// Initial tile
 		running = defkeys[*current_layout].func(&defkeys[*current_layout].arg);
 
 		//Main Loop
-		while (running == TRUE)
-		{
+		while (running == TRUE) {
 			winnum_start = countwindows(NOLOCK);
 
 			wakeupsigs = Wait((mainsig) | (1UL << cmo.mp->mp_SigBit));
 
-			if((wakeupsigs & mainsig) != 0UL) {
+			if ((wakeupsigs & mainsig) != 0UL) {
 				if (winfo[awin_index].wbwin == TRUE) {
 					Signal(subtask, subsig);
 					continue;
@@ -614,37 +457,34 @@ short int commo(void)
 				Signal(subtask, subsig);
 			}
 
-			while ((msg = (void *)GetMsg(cmo.mp)) != NULL)
-			{
+			while ((msg = (void *)GetMsg(cmo.mp)) != NULL) {
 				long id = CxMsgID(msg);
 				unsigned long type = CxMsgType(msg);
 
 				ReplyMsg((struct Message *)msg);
 
-				if (type == (unsigned long)CXM_COMMAND)
-				{
-					switch (id)
-					{
-						case CXCMD_UNIQUE:
-							if(info_on == TRUE) {
-								info_window(warn_messages[UQWARN]);
-							}
-							running = FALSE;
-							break;
-						case CXCMD_DISABLE:
-						case CXCMD_ENABLE:
-							break;
-						case CXCMD_KILL:
-							running = FALSE;
-							break;
-						case CXCMD_APPEAR:
-							break;
-						case CXCMD_DISAPPEAR:
-							break;
-						default:
-							// Do nothing
-							break;
+				if (type == (unsigned long)CXM_COMMAND) {
+					switch (id) {
+					case CXCMD_UNIQUE:
+						if (info_on == TRUE) {
+							info_window(warn_messages[UQWARN]);
 						}
+						running = FALSE;
+						break;
+					case CXCMD_DISABLE:
+					case CXCMD_ENABLE:
+						break;
+					case CXCMD_KILL:
+						running = FALSE;
+						break;
+					case CXCMD_APPEAR:
+						break;
+					case CXCMD_DISAPPEAR:
+						break;
+					default:
+						// Do nothing
+						break;
+					}
 				} else if (type == (unsigned long)CXM_IEVENT) {
 					if (id <= (TILE_FUNC_LIMIT)) {
 						*current_layout = id;
@@ -672,13 +512,13 @@ static short alloc_opts(const char *t, Ostore *s, size_t i, int subtract)
 {
 	int cxint = defopts[i].cxint - subtract;
 
-	s->strings[cxint] = malloc((strnlen(t, TT_MAX_LENGTH)+1U) * sizeof(unsigned char));
+	s->strings[cxint] = malloc((strnlen(t, TT_MAX_LENGTH) + 1U) * sizeof(unsigned char));
 
 	if (s->strings[cxint] != NULL) {
 		(void)snprintf((char *)s->strings[cxint], TT_MAX_LENGTH, "%s", t);
-		return(TRUE);
+		return (TRUE);
 	} else {
-		return(FALSE);
+		return (FALSE);
 	}
 }
 
@@ -711,13 +551,12 @@ static void free_opts(void)
 
 static short assign_bar_item(Bar_Text *b, enum bar_texts x, const char *c)
 {
-        int res = snprintf((char *)b[x].text, TT_MAX_LENGTH, "%s", c);
-        if (res < 0) {
-                return FALSE;
-        }
-        return TRUE;
+	int res = snprintf((char *)b[x].text, TT_MAX_LENGTH, "%s", c);
+	if (res < 0) {
+		return FALSE;
+	}
+	return TRUE;
 }
-
 
 static void subactionchk(void)
 {
@@ -737,7 +576,7 @@ static void subactionchk(void)
 		(void)Wait((subsig));
 	}
 
-	while(running == TRUE) {
+	while (running == TRUE) {
 		time_delay(tr, &currentval);
 		int wincnt = countwindows(NOLOCK);
 		if (wincnt > (DIVISOR - 2)) {
@@ -751,16 +590,13 @@ static void subactionchk(void)
 				(void)Wait((subsig));
 			}
 
-			if (awin_comp == NULL ||
-				awin_comp != active_win ||
-				(awin_comp->Flags & (unsigned long)WFLG_WINDOWACTIVE) == 0U) {
+			if (awin_comp == NULL || awin_comp != active_win || (awin_comp->Flags & (unsigned long)WFLG_WINDOWACTIVE) == 0U) {
 				getactive();
 				awin_comp = active_win;
 				update_wbar();
 				Signal(maintask, mainsig);
 				(void)Wait((subsig));
 			}
-
 			// If previous active window is no longer active, refresh bar
 			if ((awin_comp->Flags & (unsigned long)WFLG_WINDOWACTIVE) == 0U) {
 				getactive();
@@ -796,9 +632,7 @@ struct timerequest *create_timer(unsigned long unit)
 		return NULL;
 	}
 
-	TimerIO =
-	    (struct timerequest *)CreateExtIO(timerport,
-					      sizeof(struct timerequest));
+	TimerIO = (struct timerequest *)CreateExtIO(timerport, sizeof(struct timerequest));
 	if (TimerIO == NULL) {
 		DeletePort(timerport);	/* Delete message port */
 		return NULL;
@@ -814,7 +648,8 @@ struct timerequest *create_timer(unsigned long unit)
 	return TimerIO;
 }
 
-void time_delay(struct timerequest *tr, const struct timeval *tv) {
+void time_delay(struct timerequest *tr, const struct timeval *tv)
+{
 	unsigned short traddreq = TR_ADDREQUEST;
 	tr->tr_node.io_Command = traddreq;	/* add a new timer request */
 
@@ -825,7 +660,8 @@ void time_delay(struct timerequest *tr, const struct timeval *tv) {
 	(void)DoIO((struct IORequest *)tr);
 }
 
-void delete_timer(struct timerequest *tr) {
+void delete_timer(struct timerequest *tr)
+{
 	if (tr != 0) {
 		struct MsgPort *tp;
 		tp = tr->tr_node.io_Message.mn_ReplyPort;
@@ -835,9 +671,9 @@ void delete_timer(struct timerequest *tr) {
 		}
 
 		if (!(CheckIO((struct IORequest *)tr))) {
-			AbortIO((struct IORequest *)tr);    /* Ask device to abort any pending requests */
+			AbortIO((struct IORequest *)tr);	/* Ask device to abort any pending requests */
 		}
-		(void)WaitIO((struct IORequest *)tr);         /* Clean up */
+		(void)WaitIO((struct IORequest *)tr);	/* Clean up */
 		CloseDevice((struct IORequest *)tr);
 		DeleteExtIO((struct IORequest *)tr);
 	}
@@ -881,9 +717,9 @@ static void cleanup(struct Cmo *cmo)
 		if (wbw != NULL) {
 			CloseWindow(wbw);
 		}
-        	for (int i = 0; i < BAR_LAST_TEXT; i++) {
-                	free(bar_text[i].text);
-        	}
-        	free(bar_text);
+		for (int i = 0; i < BAR_LAST_TEXT; i++) {
+			free(bar_text[i].text);
+		}
+		free(bar_text);
 	}
 }
